// packages/flowagent-design/packages/upload/src/interface.tsx
export const STATUS: {
  [key: string]: UploadStatus
} = {
  INIT: "init",
  UPLOADING: "uploading",
  SUCCESS: "done",
  FAIL: "error",
}

export type UploadStatus = "init" | "uploading" | "done" | "error"

export type CustomIconType = {
  previewIcon?: ReactNode
  removeIcon?: ReactNode
  fileIcon?: ReactNode
  reuploadIcon?: ReactNode
  cancelIcon?: ReactNode
  startIcon?: ReactNode
  successIcon?: ReactNode
  fileName?: (file: UploadItem) => ReactNode
  progressRender?: (
    file: UploadItem,
    originDom: ReactNode,
  ) => React.ReactElement
  imageRender?: (file: UploadItem) => React.ReactNode
}
// ...
export interface UploadListProps {
  listType?: string
  fileList?: UploadItem[]
  showUploadList?: boolean | CustomIconType
  onUpload?: (file: UploadItem) => void
  onRemove?: (file: UploadItem) => void
  onReupload?: (file: UploadItem) => void
  onPreview?: (file: UploadItem) => void
  disabled?: boolean
  renderUploadItem?: (
    originNode: ReactNode,
    file: UploadItem,
    fileList: UploadItem[],
  ) => ReactNode
  renderUploadList?: (
    fileList: UploadItem[],
    uploadListProps: Omit<UploadListProps, "renderUploadList">,
  ) => ReactNode
  progressProps?: Partial<ProgressProps>
}

export type UploadItem = {
  uid: string
  status?: UploadStatus
  originFile?: File
  percent?: number
  response?: object
  url?: string
  name?: string
  children?: ReactNode
}
// ...
export interface UploadProgressProps extends CustomIconType {
  listType?: UploadListProps["listType"]
  file: UploadItem
  progressProps?: Partial<ProgressProps>
  onReupload?: UploadListProps["onReupload"]
  onUpload?: UploadListProps["onUpload"]
}